requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
cloudinary==1.30.0
Django==4.1.2
django-cloudinary-storage==0.3.0
djangorestframework==3.14.0
gunicorn==20.1.0
heroku==0.1.4
idna==3.4
Pillow==9.2.0
python-dateutil==1.5
pytz==2022.5
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
urllib3==1.26.12
whitenoise==6.2.0
